refactor: executorSchema utils to own folder
...and separate out a few methods from toExecutorSchema
