+++ .c8rc.json
@@ -4,7 +4,7 @@
   "exclude": [
     "src/**/index.ts",
     "src/**/*-fuzz.ts",
-    "src/execution/executorSchema.ts",
+    "src/executorSchema/executorSchema.ts",
     "src/overrides.d.ts",
     "src/jsutils/Maybe.ts",
     "src/jsutils/ObjMap.ts",
