+++ src/execution/index.ts
@@ -2,13 +2,6 @@ export { pathToArray as responsePathAsArray } from '../jsutils/Path';
 
 export type { ExecutionArgs } from './execute';
 
-export type {
-  ExecutorSchema,
-  GraphQLNullableInputType,
-  GraphQLNullableOutputType,
-} from './executorSchema';
-export { toExecutorSchema } from './toExecutorSchema';
-
 export {
   Executor,
   defaultFieldResolver,
