+++ src/executorSchema/index.ts
@@ -1,0 +1,6 @@
+export type {
+  ExecutorSchema,
+  GraphQLNullableInputType,
+  GraphQLNullableOutputType,
+} from './executorSchema';
+export { toExecutorSchema } from './toExecutorSchema';
